requirements.txt
azure-functions
azure-functions-durable
requests
pytz
py3-validate-email
mailjet_rest
